requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-openai
python-dotenv
pydantic-settings
streamlit
